refactor: move 'start' and 'end' events to the transaction runner
