+++ lib/Dredd.js
@@ -11,7 +11,7 @@ const handleRuntimeProblems = require('./handleRuntimeProblems');
 const resolveLocations = require('./resolveLocations');
 const resolveModule = require('./resolveModule');
 const logger = require('./logger');
-const Runner = require('./TransactionRunner');
+const TransactionRunner = require('./TransactionRunner');
 const { applyConfiguration } = require('./configuration');
 
 
@@ -33,13 +33,13 @@ class Dredd {
     };
     this.files = [];
     this.transactions = [];
-    this.runner = new Runner(this.configuration);
+    this.transactionRunner = new TransactionRunner(this.configuration);
     this.logger = logger;
   }
 
   run(callback) {
     this.logger.debug('Configuring reporters');
-    configureReporters(this.configuration, this.stats, this.runner);
+    configureReporters(this.configuration, this.stats, this.transactionRunner);
 
     this.logger.debug('Resolving --require');
     if (this.configuration.options.require) {
@@ -62,7 +62,6 @@ class Dredd {
       return;
     }
 
-    // Spin that merry-go-round
     this.logger.debug('Reading API description files.');
     this.loadFiles((loadErr) => {
       if (loadErr) { return callback(loadErr, this.stats); }
@@ -71,17 +70,9 @@ class Dredd {
       this.compileTransactions((compileErr) => {
         if (compileErr) { return callback(compileErr, this.stats); }
 
-        this.logger.debug('Starting reporters and waiting until all of them are ready.');
-        this.emitStart((emitStartErr) => {
-          if (emitStartErr) { return callback(emitStartErr, this.stats); }
-
-          this.logger.debug('Starting transaction runner.');
-          this.startRunner((runnerErr) => {
-            if (runnerErr) { return callback(runnerErr, this.stats); }
-
-            this.logger.debug('Wrapping up testing.');
-            this.transactionsComplete(callback);
-          });
+        this.logger.debug('Starting transaction runner');
+        this.startTransactionRunner((runnerErr) => {
+          callback(runnerErr, this.stats);
         });
       });
     });
@@ -183,38 +174,9 @@ Is the provided path correct?
     });
   }
 
-  // Start the runner
-  emitStart(callback) {
-    // More than one reporter is supported
-    let reporterCount = this.configuration.emitter.listeners('start').length;
-
-    // When event 'start' is emitted, function in callback is executed for each
-    // reporter registered by listeners
-    this.configuration.emitter.emit('start', this.configuration.apiDescriptions, (reporterError) => {
-      if (reporterError) { this.logger.error(reporterError.message); }
-
-      // Last called reporter callback function starts the runner
-      reporterCount--;
-      if (reporterCount === 0) {
-        callback(null, this.stats);
-      }
-    });
-  }
-
-  startRunner(callback) {
-    // Run all transactions
-    this.runner.config(this.configuration);
-    this.runner.run(this.transactions, callback);
-  }
-
-  transactionsComplete(callback) {
-    let reporterCount = this.configuration.emitter.listeners('end').length;
-    this.configuration.emitter.emit('end', () => {
-      reporterCount--;
-      if (reporterCount === 0) {
-        callback(null, this.stats);
-      }
-    });
+  startTransactionRunner(callback) {
+    this.transactionRunner.config(this.configuration);
+    this.transactionRunner.run(this.transactions, callback);
   }
 }
