+++ lib/TransactionRunner.js
@@ -42,19 +42,43 @@ class TransactionRunner {
   }
 
   run(transactions, callback) {
-    logger.debug('Sorting HTTP transactions');
-    transactions = this.configuration.options.sorted ? sortTransactions(transactions) : transactions;
+    logger.debug('Starting reporters and waiting until all of them are ready');
+    this.emitStart((emitStartErr) => {
+      if (emitStartErr) { return callback(emitStartErr); }
 
-    logger.debug('Configuring HTTP transactions');
-    transactions = transactions.map(this.configureTransaction.bind(this));
+      logger.debug('Sorting HTTP transactions');
+      transactions = this.configuration.options.sorted ? sortTransactions(transactions) : transactions;
 
-    // Remainings of functional approach, probs to be eradicated
-    logger.debug('Reading hook files and registering hooks');
-    addHooks(this, transactions, (addHooksError) => {
-      if (addHooksError) { return callback(addHooksError); }
+      logger.debug('Configuring HTTP transactions');
+      transactions = transactions.map(this.configureTransaction.bind(this));
 
-      logger.debug('Executing HTTP transactions');
-      this.executeAllTransactions(transactions, this.hooks, callback);
+      logger.debug('Reading hook files and registering hooks');
+      addHooks(this, transactions, (addHooksError) => {
+        if (addHooksError) { return callback(addHooksError); }
+
+        logger.debug('Executing HTTP transactions');
+        this.executeAllTransactions(transactions, this.hooks, (execAllTransErr) => {
+          if (execAllTransErr) { return callback(execAllTransErr); }
+
+          logger.debug('Wrapping up testing and waiting until all reporters are done');
+          this.emitEnd(callback);
+        });
+      });
+    });
+  }
+
+  emitStart(callback) {
+    // More than one reporter is supported
+    let reporterCount = this.configuration.emitter.listeners('start').length;
+
+    // When event 'start' is emitted, function in callback is executed for each
+    // reporter registered by listeners
+    this.configuration.emitter.emit('start', this.configuration.apiDescriptions, (reporterError) => {
+      if (reporterError) { logger.error(reporterError.message); }
+
+      // Last called reporter callback function starts the runner
+      reporterCount--;
+      if (reporterCount === 0) { callback(); }
     });
   }
 
@@ -597,6 +621,14 @@ include a message body: https://tools.ietf.org/html/rfc7231#section-6.3\
         transaction.test = test;
         callback();
       });
+    });
+  }
+
+  emitEnd(callback) {
+    let reporterCount = this.configuration.emitter.listeners('end').length;
+    this.configuration.emitter.emit('end', () => {
+      reporterCount--;
+      if (reporterCount === 0) { callback(); }
     });
   }
 }
